feat: add TripNotFound component
