+++ lib/components/user/monitored-trip/saved-trip-screen.js
@@ -22,6 +22,7 @@ import withLoggedInUserSupport from '../with-logged-in-user-support'
 
 import SavedTripEditor from './saved-trip-editor'
 import TripBasicsPane from './trip-basics-pane'
+import TripNotFound from './trip-not-found'
 import TripNotificationsPane from './trip-notifications-pane'
 import TripSummaryPane from './trip-summary-pane'
 
@@ -163,13 +164,22 @@ class SavedTripScreen extends Component {
   }
 
   render() {
-    const { isCreating, loggedInUser, monitoredTrips, pending } = this.props
+    const {
+      intl,
+      isCreating,
+      itinerary,
+      loggedInUser,
+      monitoredTrips,
+      pending
+    } = this.props
     const isAwaiting = !monitoredTrips || (isCreating && pending)
 
     let screenContents
     if (isAwaiting) {
       // Flash an indication while the selected and saved user trips are being loaded.
       screenContents = <AwaitingScreen />
+    } else if (!itinerary) {
+      screenContents = <TripNotFound intl={intl} />
     } else {
       const monitoredTrip = this._getTripToEdit()
       const otherTripNames = monitoredTrips
